Started moving towards typescript
