+++ src/js/versionChecker.ts
